+++ Sources/KeyboardKit/Dictation/DictationContext.swift
@@ -22,9 +22,9 @@ import SwiftUI
 /// that can be used to configure the behavior and presented
 /// to users in e.g. a settings screen.
 ///
-/// KeyboardKit will automatically setup an instance of this
-/// class in ``KeyboardInputViewController/state``, then use
-/// it as global state and inject it as an environment value.
+/// KeyboardKit set up an instance of this class and injects
+/// it as an environment value when you set up your main app
+/// and keyboard as described in <doc:Getting-Started>.
 public class DictationContext: ObservableObject {
 
     /// Create a dictation context instance.