+++ Sources/KeyboardKit/Localization/Views/Locale+ListItem.swift
@@ -17,23 +17,33 @@ public extension Locale {
             locale: Locale,
             displayLocale: Locale = .current,
             dragHandle: Bool = false,
-            subtitle: String? = nil
+            subtitle: String? = nil,
+            layoutTypeDescription: String? = nil
         ) {
             self.locale = locale
             self.displayLocale = displayLocale
             self.dragHandle = dragHandle
             self.subtitle = subtitle
+            self.layoutTypeDescription = layoutTypeDescription ?? ""
         }
 
         private let locale: Locale
         private let displayLocale: Locale
         private let dragHandle: Bool
         private let subtitle: String?
+        private let layoutTypeDescription: String
 
         public var body: some View {
             Label {
                 HStack {
-                    Text(locale.localizedName(in: displayLocale) ?? "-")
+                    VStack(alignment: .leading) {
+                        Text(locale.localizedName(in: displayLocale) ?? "-")
+                        if !layoutTypeDescription.isEmpty {
+                            Text(layoutTypeDescription)
+                                .font(.caption)
+                                .foregroundColor(.secondary)
+                        }
+                    }
                     Spacer()
                     if dragHandle {
                         ListHandle()
@@ -66,9 +76,26 @@ extension Locale {
 #Preview {
     
     List {
-        Locale.ListItem(locale: .swedish)
-        Locale.ListItem(locale: .danish, displayLocale: .german)
-        Locale.ListItem(locale: .danish, displayLocale: .german, dragHandle: false, subtitle: "foo")
-        Locale.ListItem(locale: .danish, displayLocale: .german, dragHandle: true, subtitle: "foo")
+        Locale.ListItem(
+            locale: .swedish
+        )
+        Locale.ListItem(
+            locale: .danish,
+            displayLocale: .german
+        )
+        Locale.ListItem(
+            locale: .danish,
+            displayLocale: .german,
+            dragHandle: false,
+            subtitle: "foo",
+            layoutTypeDescription: "QWERTY"
+        )
+        Locale.ListItem(
+            locale: .danish,
+            displayLocale: .german,
+            dragHandle: true,
+            subtitle: "foo",
+            layoutTypeDescription: "QWERTY"
+        )
     }
 }