+++ Sources/KeyboardKit/_Deprecated/Keyboard+Deprecated.swift
@@ -41,3 +41,12 @@ public extension Keyboard.SpaceLongPressBehavior {
         }
     }
 }
+
+public extension KeyboardContext {
+    
+    @available(*, deprecated, message: "Use `settings.spaceLongPressBehavior` instead.")
+    var spaceLongPressBehavior: Keyboard.SpaceLongPressBehavior {
+        get { settings.spaceLongPressBehavior }
+        set { settings.spaceLongPressBehavior = newValue }
+    }
+}