+++ Sources/KeyboardKit/_Keyboard/Models/Keyboard+SpaceLongPressBehavior.swift
@@ -51,4 +51,3 @@ public extension EnvironmentValues {
     /// ``KeyboardSettings/spaceLongPressBehavior``.
     @Entry var keyboardSpaceLongPressBehavior: Keyboard.SpaceLongPressBehavior?
 }
-