+++ Tests/KeyboardKitTests/Layout/InputSet+ItemRowTests.swift
@@ -1,9 +1,9 @@
 //
-//  InputSet+RowTests.swift
+//  InputSet+ItemRowTests.swift
 //  KeyboardKit
 //
 //  Created by Daniel Saidi on 2020-07-03.
-//  Copyright © 2021-2025 Daniel Saidi. All rights reserved.
+//  Copyright © 2020-2025 Daniel Saidi. All rights reserved.
 //
 
 import KeyboardKit